Fix circular import errors. KC-310
