+++ unit-tests/test_command_utils.py
@@ -1,4 +1,4 @@
-from unittest import TestCase, mock
+from unittest import TestCase, mock, skip
 
 import datetime
